Link reserved fields from device metadata
