+++ Data/Engine/Containers/webui-frontend/data/web-interface/src/Devices/Tabs/Device_Metadata.jsx
@@ -3,6 +3,7 @@ import { Alert, Box, LinearProgress, Stack, TextField, Typography } from "@mui/m
 import { AgGridReact } from "ag-grid-react";
 
 import { DEVICE_DETAILS_GRID_THEME, GridShell, DEVICE_GRID_STYLE, gridFontFamily } from "./Shared.jsx";
+import { FieldNumberCellRenderer } from "../../Metadata_Field_Cells.jsx";
 import { useAppNotifications } from "../../app/hooks/useAppNotifications.js";
 import { fetchRouteJson, getRouteErrorMessage } from "../../app/routes/routeData.js";
 
@@ -118,6 +119,47 @@ function MetadataValueCell(props) {
   );
 }
 
+export function buildDeviceMetadataColumnDefs({ saveValue, valueLimit }) {
+  return [
+    {
+      headerName: "Field Number",
+      field: "default_label",
+      width: 150,
+      pinned: "left",
+      cellRenderer: FieldNumberCellRenderer,
+    },
+    {
+      headerName: "Field Description",
+      field: "label",
+      minWidth: 220,
+      flex: 1,
+      valueGetter: (params) => params.data?.description || params.data?.default_label || "",
+    },
+    {
+      headerName: "Value",
+      field: "value",
+      minWidth: 320,
+      flex: 1.8,
+      cellRenderer: MetadataValueCell,
+      cellRendererParams: { onSaveValue: saveValue, valueLimit },
+    },
+    {
+      headerName: "Modified",
+      field: "modified_at",
+      minWidth: 180,
+      flex: 0.8,
+      valueFormatter: (params) => formatTimestamp(params.value),
+    },
+    {
+      headerName: "Source",
+      field: "source",
+      minWidth: 120,
+      flex: 0.6,
+      valueFormatter: (params) => params.value || "",
+    },
+  ];
+}
+
 export default function DeviceMetadata({ deviceId, deviceGuid, hostname }) {
   const targetId = String(deviceGuid || deviceId || hostname || "").trim();
   const [rows, setRows] = useState([]);
@@ -197,43 +239,7 @@ export default function DeviceMetadata({ deviceId, deviceGuid, hostname }) {
   );
 
   const columnDefs = useMemo(
-    () => [
-      {
-        headerName: "Field Number",
-        field: "default_label",
-        width: 150,
-        pinned: "left",
-      },
-      {
-        headerName: "Field Description",
-        field: "label",
-        minWidth: 220,
-        flex: 1,
-        valueGetter: (params) => params.data?.description || params.data?.default_label || "",
-      },
-      {
-        headerName: "Value",
-        field: "value",
-        minWidth: 320,
-        flex: 1.8,
-        cellRenderer: MetadataValueCell,
-        cellRendererParams: { onSaveValue: saveValue, valueLimit },
-      },
-      {
-        headerName: "Modified",
-        field: "modified_at",
-        minWidth: 180,
-        flex: 0.8,
-        valueFormatter: (params) => formatTimestamp(params.value),
-      },
-      {
-        headerName: "Source",
-        field: "source",
-        minWidth: 120,
-        flex: 0.6,
-        valueFormatter: (params) => params.value || "",
-      },
-    ],
+    () => buildDeviceMetadataColumnDefs({ saveValue, valueLimit }),
     [saveValue, valueLimit]
   );
